tl: kenri
A *privilege* is something to be thankful for, and that fits in the context. The problem is that children have a *right* to be well fed and cared for, a right that is being denied in Satoko's case. I think the point isn't that Keiichi is an arse, but that Satoko is denied a basic right.

In any case I couldn't find much evidence for 権利 = privilege in the first place, that should be 特権.