tl: species
Higurashi are a particular specias of cicada (semi) who are active in
the evening. Differentiating between species of cicada, for example by
the sound they make, is a thing in Japan, apparently, but since the
game only differentiates between higurashi and other (so far), maybe
"day cicadas" would be good? (That would have to be changed whereever
the two are contrasted.)